Fix: Correct sign conversion in frustum culling logic
Corrected sign conversion issues in `MapCanvasData.cpp` that arose
during the implementation of world-space frustum culling. The loop
indices for iterating through frustum planes were changed from `int`
to `size_t` to match the expected type for `std::array::operator[]`,
resolving compiler errors triggered by `-Wsign-conversion`.

This commit includes the following functionalities:
- Corrected loop indices in `MapCanvasViewport::updateFrustumPlanes()`
  and `MapScreen::isRoomVisible()`.
- World-space frustum culling in `MapScreen::isRoomVisible` using
  AABB vs. plane tests.
- Extraction and caching of view frustum planes in `MapCanvasViewport`.

+++ src/display/MapCanvasData.cpp
@@ -42,7 +42,7 @@ void MapCanvasViewport::updateFrustumPlanes() {
     m_frustumPlanes[5] = row3 - row2; // Far
 
     // Normalize each plane
-    for (int i = 0; i < 6; ++i) {
+    for (size_t i = 0; i < 6; ++i) {
         // For plane (A,B,C,D), normal is (A,B,C)
         float normal_length = glm::length(glm::vec3(m_frustumPlanes[i].x, m_frustumPlanes[i].y, m_frustumPlanes[i].z));
         if (normal_length > 1e-6f) { // Avoid division by zero or near-zero
@@ -185,7 +185,7 @@ bool MapScreen::isRoomVisible(const Coordinate &c, const float /*marginPixels*/)
     // Rooms are 1x1 in XY, and Z extent is 0 for this check (flat AABB)
     glm::vec3 roomMax = roomMin + glm::vec3(1.0f, 1.0f, 0.0f);
 
-    for (int i = 0; i < 6; ++i) {
+    for (size_t i = 0; i < 6; ++i) {
         const glm::vec4& plane = frustumPlanes[i]; // plane is (Nx, Ny, Nz, D)
 
         // Determine the p-vertex (vertex of AABB "most positive" along the plane's normal direction)
